+++ process.c
@@ -66,13 +66,13 @@ Boolean setup(FilterRecordPtr pb){
 		var['Y'] = FILTER_RECT(pb).bottom - FILTER_RECT(pb).top;
 	}
 	var['Z'] = nplanes;
-	var['D'] = ff_D();
+	var['D'] = val_D;
 	var['M'] = ff_M();
 
 	var['R'] = var['G'] = var['B'] = var['A'] = var['C'] = 255;
-	var['I'] = max_channel_i;
-	var['U'] = max_channel_u;
-	var['V'] = max_channel_v;
+	var['I'] = val_I;
+	var['U'] = val_U;
+	var['V'] = val_V;
 
 	/* initialise flags for tracking special variable usage */
 	for(i = 0; i < 0x100; i++)